+++ crates/vm/levm/bench/revm_comparison/src/benchmark.rs
@@ -1,4 +1,6 @@
-use revm_comparison::{generate_calldata, load_contract_bytecode, run_with_levm, run_with_revm};
+use revm_comparison::{levm_bench::run_with_levm, revm_bench::run_with_revm};
+use sha3::{Digest, Keccak256};
+use std::{fs::File, io::Read};
 
 enum VM {
     Revm,
@@ -42,4 +44,40 @@ fn main() {
         VM::Levm => run_with_levm(&bytecode, runs, &calldata),
         VM::Revm => run_with_revm(&bytecode, runs, &calldata),
     }
+}
+
+// Auxiliary functions for getting calldata and bytecode.
+
+fn generate_calldata(function: &str, n: u64) -> String {
+    let function_signature = format!("{}(uint256)", function);
+    let hash = Keccak256::digest(function_signature.as_bytes());
+    let function_selector = &hash[..4];
+
+    // Encode argument n (uint256, padded to 32 bytes)
+    let mut encoded_n = [0u8; 32];
+    encoded_n[24..].copy_from_slice(&n.to_be_bytes());
+
+    // Combine the function selector and the encoded argument
+    let calldata: Vec<u8> = function_selector
+        .iter()
+        .chain(encoded_n.iter())
+        .copied()
+        .collect();
+
+    hex::encode(calldata)
+}
+
+fn load_contract_bytecode(bench_name: &str) -> String {
+    let path = format!(
+        "{}/contracts/bin/{}.bin-runtime",
+        env!("CARGO_MANIFEST_DIR"),
+        bench_name
+    );
+
+    println!("Loading bytecode from file {}", path);
+
+    let mut file = File::open(path).unwrap();
+    let mut contents = String::new();
+    file.read_to_string(&mut contents).unwrap();
+    contents
 }
